feat: 이미지 관련 기능 수정 및 추가

+++ backend/src/main/java/com/project/vegan/domain/diet/service/DietService.java
@@ -99,7 +99,9 @@ public class DietService {
             throw new ForbiddenException();
         }
 
-        dietUploadFileRepository.deleteAll(dietUploadFileRepository.findByDiet(diet));
+        if(dietUploadFileRepository.existsByDiet(diet)){
+            dietUploadFileRepository.deleteAll(dietUploadFileRepository.findByDiet(diet));
+        }
 
         dietRepository.delete(diet);
     }
